Move remaining status related logic to Status column component.
